teaspoons-cli/pyproject.toml
[tool.poetry]
name = "teaspoons-cli"
version = "0.0.1"
description = "Command line interface for interacting with the Terra Scientific Pipelines Service, or Teaspoons."
authors = ["Terra Scientific Services <teaspoons-developers@broadinstitute.org>"]
readme = "README.md"
packages = [
    {include = "teaspoons"}
]

[tool.poetry.dependencies]
python = "^3.9"
teaspoons-client = {path = "generated"}
python-dotenv = "^1.0.1"
click = "^8.1.7"
oauth2-cli-auth = "^1.5.0"
PyJWT = "^2.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^3.0.0"
mockito = ">=1.4.0"
pytest-mockito = ">=0.0.4"
ruff = "^0.7.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "teaspoons"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "generated"
]
# Same as Black.
line-length = 88
indent-width = 4
